added some physics and a few menu changes
added friction, acceleration, and gravity. Though he does collide with the floor, mario just accelerates right through the floor. also added a few of the animations, and displayed numbers for score, coin count, and time, as well as the high score on the title screen. additionally, you cant move him up or down though the print statements do say that the key presses work. probably something to do with the physics changes i've made so far but i gotta check on that later